refactor: standardize vault key API endpoints and actions
