+++ packages/insomnia/src/ui/index.tsx
@@ -1257,24 +1257,24 @@ async function renderApp() {
                 element: <Authorize />,
               },
               {
-                path: 'updateVaultSalt',
-                action: async (...args) => (await import('./routes/auth.vaultKey')).updateVaultSaltAction(...args),
+                path: 'update-vault-salt',
+                action: async args => (await import('./routes/auth.update-vault-salt')).action(args),
               },
               {
-                path: 'createVaultKey',
-                action: async (...args) => (await import('./routes/auth.vaultKey')).createVaultKeyAction(...args),
+                path: 'create-vault-key',
+                action: async args => (await import('./routes/auth.create-vault-key')).action(args),
               },
               {
-                path: 'validateVaultKey',
-                action: async (...args) => (await import('./routes/auth.vaultKey')).validateVaultKeyAction(...args),
+                path: 'validate-vault-key',
+                action: async args => (await import('./routes/auth.validate-vault-key')).action(args),
               },
               {
-                path: 'resetVaultKey',
-                action: async (...args) => (await import('./routes/auth.vaultKey')).resetVaultKeyAction(...args),
+                path: 'reset-vault-key',
+                action: async args => (await import('./routes/auth.reset-vault-key')).action(args),
               },
               {
-                path: 'clearVaultKey',
-                action: async (...args) => (await import('./routes/auth.vaultKey')).clearVaultKeyAction(...args),
+                path: 'clear-vault-key',
+                action: async args => (await import('./routes/auth.clear-vault-key')).action(args),
               },
             ],
           },
