+++ packages/insomnia/src/ui/routes/auth.create-vault-key.tsx
@@ -1,0 +1,7 @@
+import type { ActionFunctionArgs } from 'react-router';
+
+import { createVaultKey } from '../vault-key';
+
+export async function action(_args: ActionFunctionArgs) {
+  return createVaultKey('create');
+}
